Run ethernet over USB device interface
This lets us interact with the Beaglebone and push software
updates through USB. The idea is to avoid requiring that
the Ethernet port be attached to the development PC. While
this is likely the case for now, it seems nice to be able
to do all dev on the Beaglebone through the USB cable. Also,
a couple of my "real" projects don't have Ethernet interfaces,
so this supports those use cases.

The USB ethernet interface is brought up by S06usbether. The setup
expects that the other end of the USB connection either runs
a DHCP server or support IPv4 link local addressing (zeroconf). I
currently run with the link local addressing and this seems like
the simplest way to go. The beaglebone shows up as beagleboard.local
by default and advertises the ssh service.

Signed-off-by: Frank Hunleth <fhunleth@troodon-software.com>